+++ modules/tactile/editor/gui/widgets/dialogs/resize_map_dialog.hpp
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <tactile-base/tactile_std.hpp>
+
 #include <entt.hpp>  // dispatcher
 
 namespace Tactile {
 
 void UpdateResizeMapDialog(entt::dispatcher& dispatcher);
 
-void OpenResizeMapDialog(int currentRows, int currentColumns);
+void OpenResizeMapDialog(usize currentRows, usize currentColumns);
 
 }  // namespace Tactile
